Move state modifying apis to transaction apis, update minimally required apis.
